misc: enclosure: Introduce get and set pattern.
New API will be pattern oriented, not LED oriented. The set of
possible patterns is defined by enum enclosure_led_pattern. With this
approach, the enclosure interface is not designed to particular standard
like 2 (active, fault) or 3 LED systems (active, fault and locate).
The enum will be interpreted by a driver. The interface is ready to
support enclosures where patterns are not mutally exclusive.

Enum enclosure_led_pattern is based on IBPI standard[1] and NPEM[2].
ENCLOSURE_LED_ACTIVE is not defined on purpose. In general, activity
control on the drive should belongs to hardware, not to userspace.

Enum enclosure_status is used as a return value for set method. Even if
it is ses-2 specyfic, those values are self describing enough to be
reused.
New API consumer will be added in next patches.

[1] https://www.snia.org/node/4691
[2] PCie Base Specification r6.0.1-1.0 sec 6.28

Signed-off-by: Mariusz Tkaczyk <mariusz.tkaczyk@linux.intel.com>