pci: Register Native PCIe Enclosure Management
NPEM is PCIe extended capability so it should be registered in
pcie_init_capabilities() but it is not possible due to led dependency.
Parent pci_device must be added earlier for led_classdev_register()
to be successful. NPEM does not require configuration on kernel side, it
is safe to register led devices later.

Signed-off-by: Mariusz Tkaczyk <mariusz.tkaczyk@linux.intel.com>