pcie: NPEM support
NPEM is a pattern oriented and it does not give a direct access to LED.
To fit it in leds subsystem, driver creates a separate ledclass_dev for
every supported pattern. For every led only 2 brightness states
are allowed: LED_ON or LED_OFF. As a result, one LED entry represents
one pattern.

Definition of patterns are described in PCIE Base Specification r6.1 [1].
The interface is ready to support enclosures where patterns are not
mutually exclusive, driver may clear other LEDs if they cannot be enabled
together.

Driver is projected to be exclusive NPEM extended capability manager.
It waits up to 1 second after imposing new request, it doesn't verify if
controller is busy before write, assuming that mutex lock gives protection
from concurrent updates. Driver is not be registered if _DSM led management
is available.

Link: https://members.pcisig.com/wg/PCI-SIG/document/19849 [1]
Link: https://lore.kernel.org/linux-pci/cover.1643822289.git.stuart.w.hayes@gmail.com/ [2]
Link: https://lore.kernel.org/all/20210601203820.3647-1-stuart.w.hayes@gmail.com/ [3]

Signed-off-by: Mariusz Tkaczyk <mariusz.tkaczyk@linux.intel.com>